moved test to correct file and made the CI command more specific
