+++ .github/workflows/continuous-integration.yml
@@ -26,5 +26,5 @@ jobs:
       # run unittests
       - name: Run unittests
         run: |
-          python -m unittest test
+          python -m unittest discover -s test
